requirements.txt
requests
python-dotenv
cohere